Add OverrideReference.
