+++ tutor_recon/config/__init__.py
@@ -1,6 +1,7 @@
 from . import (
     configs,
-    main,
+    override_sequence,
+    reference,
     override,
     templates,
     tutor,
